feat(entries): 添加删除记录功能并优化勾选状态处理
- 新增删除记录 API 接口 deleteEntry
- 将勾选状态类型从字符串改为数字类型
- 优化记录勾选状态更新逻辑，支持乐观更新回滚
- 移除 DailyEntries 组件中的乐观更新相关方法
- 修改 RecordPage 页面记录添加逻辑，改为添加后刷新列表
- 更新 EntriesUpdateCheckedReq 请求参数校验注解
- 优化数据库查询 SQL 语法格式
- 添加新的构建脚本 pure-build
- 更新应用标题为 Chronicle
- 调整环境配置文件中的 API 地址设置

+++ src/main/java/com/cxin/chronicle/infrastructure/model/entity/Entries.java
@@ -52,7 +52,7 @@ public class Entries implements Serializable {
      * 是否勾选
      */
     @TableField("checked")
-    private Byte checked;
+    private int checked;
 
     /**
      * 创建时间
